# contracts/openapi/jk-platform.yaml
...
openapi: 3.1.0
info:
  title: SAGA API
  version: 1.0.0
  description: Command and integration API for SAGA.
servers:
  - url: https://api.example.com/api/v1
security:
  - oidc: []
...
paths:
...
  /animals:
    post:
      operationId: registerAnimal
      summary: Register an animal
      tags: [Animals]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterAnimalRequest" }
      responses:
        "201":
          description: Animal registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Animal" }
        "409": { $ref: "#/components/responses/Conflict" }
    get:
      operationId: listAnimals
      summary: List animals in the tenant (optionally by farm)
      tags: [Animals]
      parameters:
        - name: farmId
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Animal list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Animal" }
  /animals/{animalId}:
    get:
      operationId: getAnimal
      tags: [Animals]
      parameters:
        - $ref: "#/components/parameters/AnimalId"
      responses:
        "200":
          description: Authorized animal summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Animal" }
        "404": { $ref: "#/components/responses/NotFound" }
  /animals/{animalId}/photos:
    post:
      operationId: uploadAnimalPhoto
      summary: Upload a dated animal photo (JK-ANI photo gallery)
      tags: [Animals]
      parameters:
        - $ref: "#/components/parameters/AnimalId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, takenAt]
              properties:
                file: { type: string, format: binary }
                takenAt: { type: string, format: date }
                caption: { type: string, maxLength: 500 }
      responses:
        "201":
          description: Photo recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnimalPhoto" }
        "415": { description: Unsupported content type }
        "413": { description: File too large }
    get:
      operationId: listAnimalPhotos
      summary: Chronological photo gallery, newest first (active only)
      tags: [Animals]
      parameters:
        - $ref: "#/components/parameters/AnimalId"
      responses:
        "200":
          description: Photo gallery
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AnimalPhoto" }
  /animals/{animalId}/photos/{photoId}/download:
    get:
      operationId: downloadAnimalPhoto
      summary: Stream a photo's bytes
      tags: [Animals]
      parameters:
        - $ref: "#/components/parameters/AnimalId"
        - name: photoId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Photo bytes
          content:
            image/jpeg: { schema: { type: string, format: binary } }
            image/png: { schema: { type: string, format: binary } }
            image/webp: { schema: { type: string, format: binary } }
        "404": { $ref: "#/components/responses/NotFound" }
  /animals/{animalId}/photos/{photoId}:
    delete:
      operationId: removeAnimalPhoto
      summary: Soft-remove a photo (compensating event, not a hard delete)
      tags: [Animals]
      parameters:
        - $ref: "#/components/parameters/AnimalId"
        - name: photoId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, maxLength: 500 }
      responses:
        "204": { description: Photo removed }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /animals/{animalId}/timeline:
    get:
      operationId: getAnimalTimeline
      summary: Cursor/limited chronological event timeline (JK-ANI-004)
      tags: [Animals]
      parameters:
        - $ref: "#/components/parameters/AnimalId"
        - {
            name: limit,
            in: query,
            required: false,
            schema: { type: integer, minimum: 1, maximum: 500 },
          }
        - { name: types, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Timeline entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { type: object } }
  /animals/{animalId}/identifiers:
    post:
      operationId: assignOrReplaceIdentifier
      summary: Assign or replace an identifier (JK-ANI-002, JK-DOM-003)
      tags: [Animals]
      parameters:
        - $ref: "#/components/parameters/AnimalId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifierType, identifierValue]
              properties:
                identifierType: { type: string, enum: [rfid, visual, official, legacy] }
                identifierValue: { type: string }
                replace: { type: boolean, default: false }
                reason: { type: string }
      responses:
        "201": { description: Identifier assigned or replaced }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example.com/.well-known/openid-configuration
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >
        Client-generated idempotency key (§46). Retried commands with the same
        key return the original result instead of duplicating the effect.
      schema: { type: string, minLength: 16, maxLength: 200 }
    AnimalId:
      name: animalId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    RegisterAnimalRequest:
      type: object
      additionalProperties: false
      required: [farmId, visualId, sex, breedCode]
      properties:
        farmId: { type: string, format: uuid }
        visualId: { type: string, minLength: 1, maxLength: 80 }
        rfid: { type: [string, "null"], maxLength: 128 }
        sex: { type: string, enum: [female, male, unknown] }
        speciesCode:
          { type: string, enum: [BOVINE, PORCINE, OVINE, CAPRINE, EQUINE], default: BOVINE }
        breedCode: { type: string, default: BRANGUS }
        birthDate: { type: [string, "null"], format: date }
    Animal:
      type: object
      required: [id, farmId, visualId, sex, breedCode, lifecycleStatus, version]
      properties:
        id: { type: string, format: uuid }
        farmId: { type: string, format: uuid }
        visualId: { type: string }
        sex: { type: string }
        speciesCode: { type: string, enum: [BOVINE, PORCINE, OVINE, CAPRINE, EQUINE] }
        breedCode: { type: string }
        birthDate: { type: [string, "null"], format: date }
        lifecycleStatus: { type: string }
        version: { type: integer, minimum: 0 }
    AnimalPhoto:
      type: object
      required:
        [
          id,
          animalId,
          takenAt,
          contentType,
          byteSize,
          checksumSha256,
          status,
          createdAt,
        ]
      properties:
        id: { type: string, format: uuid }
        animalId: { type: string, format: uuid }
        takenAt: { type: string, format: date }
        caption: { type: [string, "null"] }
        contentType: { type: string, enum: [image/jpeg, image/png, image/webp] }
        byteSize: { type: integer, minimum: 1 }
        checksumSha256: { type: string }
        status: { type: string, enum: [active, removed] }
        createdAt: { type: string, format: date-time }
...
    Problem:
      type: object
      description: RFC 9457 Problem Details with JK extensions (§46.1).
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        correlationId: { type: string, format: uuid }
        errors:
          type: array
          items:
            type: object
            required: [field, reason]
            properties:
              field: { type: string }
              reason: { type: string }
...
  responses:
    Conflict:
      description: Business conflict
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Resource not found or not visible
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...